💥 Change Size's DataType from uint32_t to uint64_t
So that the type matches the one of vector::size()

+++ src/Size.h
@@ -55,6 +55,6 @@ private:
     DataType _height = 1;
 };
 
-using Size = SizeT<uint32_t>;
+using Size = SizeT<uint64_t>;
 
 } // namespace img
